pyproject.toml
[project]
name = "my-mcp-server"
version = "0.1.0"
description = "My personal MCP server with modular architecture and example tools."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "mcp>=1.0.0",
  "httpx>=0.28.1",
  "click>=8.0.0",
  "python-dotenv>=1.0.0",
  "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.8.0",
  "pytest>=7.4.0",
  "pytest-asyncio>=0.23.0",
  "types-psutil>=5.9.0",
]

[project.scripts]
my-mcp-server = "main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]
py-modules = ["main"]

[tool.setuptools.package-dir]
"" = "."
"src" = "src"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false    # Start lenient, tighten later
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
show_column_numbers = true

# Per-module options
[[tool.mypy.overrides]]
module = "mcp.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "httpx.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "click.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "dotenv.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "psutil.*"
ignore_missing_imports = true
